Move spindle-parse-result
