+++ frontend/src/app/shared/components/spindle-parse-result/spindle-parse-result.component.html
