+++ frontend/src/app/shared/components/spindle-parse-result/spindle-parse-result.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { SpindleParseResultComponent } from "./spindle-parse-result.component";
-import { SharedModule } from "../shared.module";
+import { SharedModule } from "../../shared.module";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 
 describe("SpindleParseResultComponent", () => {
