+++ frontend/src/app/shared/components/spindle-parse-result/spindle-parse-result.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from "@angular/core";
-import { AlertService } from "../services/alert.service";
-import { AlertType } from "../components/alert/alert.component";
-import { SpindleApiService } from "../services/spindle-api.service";
+import { AlertService } from "../../services/alert.service";
+import { AlertType } from "../alert/alert.component";
+import { SpindleApiService } from "../../services/spindle-api.service";
 import { faCopy, faDownload } from "@fortawesome/free-solid-svg-icons";
 
 export interface TextOutput {
@@ -15,7 +15,7 @@ export interface TextOutput {
     styleUrl: "./spindle-parse-result.component.scss",
 })
 export class SpindleParseResultComponent {
-    @Input() textOutput: TextOutput | null = null;
+    @Input({ required: true }) textOutput: TextOutput | null = null;
 
     public faDownload = faDownload;
     public faCopy = faCopy;
@@ -23,9 +23,14 @@ export class SpindleParseResultComponent {
     constructor(
         private alertService: AlertService,
         private apiService: SpindleApiService,
-    ) {}
+    ) { }
 
-    public downloadAsFile = this.apiService.downloadAsFile;
+    public downloadAsFile(
+        textData: string,
+        extension: "tex" | "json" | "pdf",
+    ): void {
+        this.apiService.downloadAsFile(textData, extension);
+    }
 
     public copyToClipboard(text: string): void {
         navigator.clipboard
